fix(trailties): address Copilot PR #2185 review (round 3)
- Break the migration.ts ↔ actions/create-migration.ts import cycle by
  extracting migrationLookupAt/migrationExists into migration-lookup.ts;
  both files import the lookup helpers from there. migration.ts still
  re-exports them so external callers see the same surface.
- migrationTemplate now wraps the host with assigns.migrationFileName
  before invoking createMigration, so the CreateMigration action's
  existence checks can't drift from the just-computed assigns.
